+++ objectified-ui/public/WHATS_NEW.md
@@ -65,6 +65,7 @@ We continue to improve the platform based on your feedback with improvements and
   - Dragging a class to a response schema allows the class to be copied or referenced
 - OpenAPI Support:
   - Now at 100% support for class and properties definitions
+  - Improved import support for paths and securitySchemes
 - Bug fixes:
   - Fixed AI generation bug when creating $ref objects: was referring to properties, not the object itself.
   - Fixed $ref editing in classes - was opening a blank editing panel, now corrected.